rename li0
